Graphics.Curves.Attribute

Description

Write-only attributes.

Documentation

data Assignment a

Representation of an attribute update for an element of type a.

Constructors

forall f b . HasAttribute f a => (f b) := b
Set an attribute

forall f b . HasAttribute f a => (f b) :~ (b -> b)
Modify an attribute

class HasAttribute f a where

The type constructor f is such that f b is the type of names of attributes of a of type b.

Methods

modifyAttribute :: f b -> (b -> b) -> a -> a

setAttribute :: f b -> b -> a -> a

with :: a -> [Assignment a] -> a

Apply a sequence of attribute assignments to an object (applied left-to-right).