Graphics.Curves.Geometry

Description

Combinators for drawing various geometric figures.

Documentation

rectangle :: Point -> Point -> Image

Draw an axis-aligned rectangle with the given opposite corners.

type Length = Scalar

type Angle = Scalar

triangle :: Length -> Length -> Length -> Image

Draw a triangle with the given side lengths. The corner opposite the third side is located at the origin and the first side is drawn along the x-axis.

triangleA :: Length -> Length -> Angle -> Image

Draw a triangle given by two side lengths and an angle. The corner of the angle is drawn at the origin and the first side along the x-axis.

triangleAA :: Length -> Angle -> Angle -> Image

Draw a triangle given by two angles and the length of the side shared by the angles. The first angle is drawn at the origin and the side along the x-axis.

regularPoly :: Int -> Image

Draw an n-sided regular polygon centered at the origin and one corner at (1, 0).

angleArc

Arguments

:: Scalar
Radius of the arc in pixels

-> Point
A

-> Point
B

-> Point
C

-> Image

Draw an angle arc for the counter clockwise angle BAC.

labelledAngle

Arguments

:: String

-> Vec
Position of the text relative to the angle

-> Point

-> Point

-> Point

-> Image

Draw an angle arc labelled by the given string.

arrow :: Point -> Point -> Image

Draw a line segment with an arrow head.