Graphics.Curves.Math

Description

Simple two-dimensional linear algebra.

Vectors

type Scalar = Double

data Vec

Two-dimensional vectors.

Constructors

Vec

Fields

getX :: !Scalar

getY :: !Scalar

type Point = Vec

unitX :: Vec

unitX = Vec 1 0

unitY :: Vec

unitY = Vec 0 1

diag :: Scalar -> Vec

diag x = Vec x x

vmap :: (Scalar -> Scalar) -> Vec -> Vec

Apply a function to the coordinates of a vector.

vzip :: (Scalar -> Scalar -> Scalar) -> Vec -> Vec -> Vec

Point-wise lifting of an operator on coordinates.

vuncurry :: (Scalar -> Scalar -> a) -> Vec -> a

vuncurry f (Vec x y) = f x y

vcurry :: (Vec -> a) -> Scalar -> Scalar -> a

vcurry f x y = f (Vec x y)

dot :: Vec -> Vec -> Scalar

The dot product of two vectors.

rot90 :: Vec -> Vec

Rotate a vector 90 degrees counterclockwise.

norm :: Vec -> Vec

Normalize a vector.

norm v = v / abs v

angle :: Vec -> Vec -> Scalar

The counterclockwise angle between two vectors.

interpolate :: Point -> Point -> Scalar -> Point

The weighted average of two points.

interpolate p q t == (1 - t) * p + t * q

Line segments

data Segment

Constructors

Seg

Fields

getStart :: !Point

getEnd :: !Point

segmentLength :: Segment -> Scalar

The length of a segment.

segmentLength (Seg p q) = distance p q

squareSegmentLength :: Segment -> Scalar

The square length of a segment. Avoids computing a square root.

leftOf :: Point -> Segment -> Bool

Is a point to the left of a line segment, as seen from the start of the segment looking a the end?

intersectSegment :: Segment -> Segment -> Maybe Point

Compute the intersection point of two segments, if any.

intersectLine :: Segment -> Segment -> Maybe Point

Compute the intersection point of two lines, if any.

intersectLineSegment :: Segment -> Segment -> Maybe Point

Compute the intersection point of a line and a segment, if any.

Basis

data Basis

A basis for a coordinate system.

Constructors

Basis

Fields

origin :: Point

xUnit :: Point

yUnit :: Point

defaultBasis :: Basis

defaultBasis = Basis 0 unitX unitY

toBasis :: Basis -> Point -> Point

Translate a point from the defaultBasis to the given basis.

fromBasis :: Basis -> Point -> Point

Translate a point in the given basis to the defaultBasis.

Distances

class DistanceToPoint a where

Minimal complete definition

Nothing

Methods

distance :: a -> Point -> Scalar

Compute the distance from an a to a given point. Default implementation:

distance x p = sqrt (squareDistance x p)

squareDistance :: a -> Point -> Scalar

The square of the distance from an a to a point. Default implementation:

squareDistance x p = distance x p ^ 2

distanceAtMost :: Scalar -> a -> Point -> Maybe Scalar

The distance from an a to a point if it's less than a given value. distanceAtMost d x p == Nothing implies that distance x p > d.

Transformations

class Transformable a where

Methods

transform :: (Point -> Point) -> a -> a

Apply a transformation to all points in an object.

translate :: Transformable a => Vec -> a -> a

translate v = transform (+ v)

scale :: Transformable a => Vec -> a -> a

scale v = transform (* v)

scaleFrom :: Transformable a => Point -> Vec -> a -> a

Scale using a given point as the center.

rotate :: Transformable a => Scalar -> a -> a

Rotate an object counterclockwise around the origin.

rotateAround :: Transformable a => Point -> Scalar -> a -> a

Rotate an object counterclockwise around a given point.

Function analysis

findThreshold

Arguments

:: (Scalar -> a)
The function to analyze.

-> (a -> Bool)
The predicate.

-> Scalar
Precision. The actual smallest value will be less than this much smaller than the returned value.

-> Scalar
A minimum value. No solution smaller than this will be returned.

-> Scalar
A maximum value. If no solution is found below this value, Nothing is returned.

-> Maybe (Scalar, a)

Find the smallest value making a function satisfy a given predicate. Needs the function to be monotone in the predicate to work properly.