Graphics.Curves.SVG.Font

Description

This module contains functions to render text using fonts specified in the SVG format.

Documentation

data SVGFont

Contains all the data necessary to render text.

loadFont :: FilePath -> IO SVGFont

Read a font from an SVG file.

drawString :: SVGFont -> String -> Image

Render a string in the given font. The text starts at the origin and is scaled to make upper case letters 1 unit high.