Graphics.Curves.Text

Description

Simple (ugly) hand-crafted text. See Graphics.Curves.SVG.Font for prettier text.

Documentation

data Alignment

Constructors

LeftAlign

RightAlign

CenterAlign

stringImage :: String -> Image

Equivalent to stringImage' LeftAlign 0.2.

stringImage' :: Alignment -> Scalar -> String -> Image

Draw a string at the origin with the given text alignment. The second argument specifies the spacing between characters. Upper-case letters are 2 units high.

label :: Point -> Scalar -> String -> Image

Draw a string centered at a given point. The second argument specifies the font height in pixels, invariant under scaling.